requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
contourpy==1.1.1
cycler==0.12.1
efficientnet_pytorch==0.7.1
filelock==3.16.1
fonttools==4.55.2
fsspec==2024.10.0
huggingface-hub==0.26.5
idna==3.10
importlib_resources==6.4.5
joblib==1.4.2
kiwisolver==1.4.7
matplotlib==3.7.5
numpy==1.24.4
opencv-python==4.10.0.84
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
safetensors==0.4.6.dev0
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
threadpoolctl==3.5.0
timm==1.0.12
torch==1.12.1+cu116
torchvision==0.13.1+cu116
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
zipp==3.20.2
